Implement log preview
