+++ src/views/log_view.py
@@ -1,105 +1,37 @@
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 
-from src.heuristics import manager as heuristics
-from src.views import View, HeuristicSetup, SelectFiles, DrainSetup
+from nicegui.element import Element
 
+from src.logs.types import LogFile
+
+from .view import View
 from nicegui import ui
-from nicegui.tailwind_types.text_color import TextColor
 
 
 @dataclass
 class LogView(View):
-    drain_setup: DrainSetup
-    heuristic_setup: HeuristicSetup
-    select_files: SelectFiles
+    log_file: LogFile
+    lines: list[Element] = field(default_factory= lambda: [])
 
-    _drain_needs_calculation: bool = False
+    def show(self) -> Element:
+        with ui.element("div") as e:
+            e.tailwind.space_between("y-1.5").max_width("2xl")
+            for i, line in enumerate(self.log_file.lines):
+                lbl = ui.label(f"{i + 1}: {line.line}")
+                self.lines.append(lbl)
+                lbl.tailwind.font_family("mono")
+        return e
 
-    def update_log_visualization(self):
-        self.parent.tailwind.padding("p-5").container().box_shadow(
-            "inner"
-        ).background_color("zinc-300").min_height("max")
+    def focus_line(self, id: int):
+        for line in self.lines:
+            line.classes(
+                remove="font-bold"
+            )
 
-        if self.select_files.state == SelectFiles.State.FILES_NOT_UPLOADED:
-            self._show_logs_not_loaded()
-            return
-        
-        if self._drain_needs_calculation:
-            self._recalculate_drain()
-
-        self._show_logs()
-    
-    def _recalculate_drain(self):
-        assert (
-            self.select_files.grand_truth is not None
-            and self.select_files.checked is not None
+        self.lines[id].run_method(
+            "scrollIntoView", {"behavior": "smooth", "block": "center"}
         )
-
-        drain = self.drain_setup.build_drain()
-        drain.learn(self.select_files.grand_truth)
-        drain.learn(self.select_files.checked)
-        drain.annotate(self.select_files.grand_truth)
-        drain.annotate(self.select_files.checked)
-        heuristics.apply_heuristics(
-            self.select_files.grand_truth, self.select_files.checked
-        )
-        self._drain_needs_calculation = False
-        
-
-    def _show_logs_not_loaded(self):
-        ui.label("Please provide files above to see logs here !").tailwind.text_align(
-            "center"
-        ).width("full").font_size("lg")
-
-    def _show_logs(self):
-        COLORS: list[TextColor] = [
-            "neutral-500",
-            "neutral-500",
-            "green-500",
-            "green-600",
-        ]
-
-        assert self.select_files.checked is not None
-
-        for i, line in enumerate(self.select_files.checked.lines):
-            val = 0
-            for heuristic in line.list_heuristics():
-                val = max(val, line.get_heuristic(heuristic))
-
-            if val < self.heuristic_setup.heuristic_cap:
-                continue
-
-            color = COLORS[int(len(COLORS) * val)]
-
-            lbl = ui.label(f"{i + 1}: {line.line}")
-            lbl.tailwind.font_family("mono").user_select("none").text_overflow(
-                "text-clip"
-            ).text_color(color)
-            lbl.classes(add="hover:font-bold")
-            with lbl:
-                tooltip_text = [
-                    f"Template: {line.template.pattern}",
-                ]
-                tooltip_text += [
-                    f"{heuristic}: {line.get_heuristic(heuristic)}"
-                    for heuristic in line.list_heuristics()
-                ]
-                ui.tooltip("\n".join(tooltip_text)).tailwind.font_size(
-                    "base"
-                ).font_family("mono").whitespace("pre-line")
-
-    def show(self):
-        self.parent = ui.element("div")
-        self.heuristic_setup.state_changed.connect(self.update, weak=False)
-        self.select_files.state_changed.connect(self.update, weak=False)
-        self.drain_setup.state_changed.connect(self.update, weak=False)
-        with self.parent:
-            self.update_log_visualization()
+        self.lines[id].tailwind.font_weight("bold")
 
     def update(self, sender: object = None):
-        if isinstance(sender, (SelectFiles, DrainSetup)):
-            self._drain_needs_calculation = True
-
-        self.parent.clear()
-        with self.parent:
-            self.update_log_visualization()
+        pass
